Extract concurrent config and result validation into _concurrent_config
Move the ConcurrentConfig and ConcurrentResult dataclasses and their eager
validation out of cuprum/concurrent.py into a dedicated module, splitting the
failure-index and submission-index checks into focused helpers. Replace the
oversized test_concurrency module with four cohesive suites covering config,
hooks, execution, and fail-fast behaviour.
